+ make pageselector work bette, move itout of vcl table and in the query
this is not quite right though, cause if acl checks fail, we will get pages that have less objects on them than advertised, but then again, I have no idea how we could do acl checks for all 8000 objects in stat's users folder. so we only do them for the objects on the current page.