Support for loading things on-demand.
How it works:

$tree->start_tree(array(
	"type" => TREE_DHTML,
))

generates the old style tree where everything it loaded at once.


$tree->start_tree(array(
	"type" => TREE_DHTML,
	"get_branch_func" => $this->mk_my_orb("gen_folders",array("parent" => ""),"workbench"),
));

generates a tree where only the first level is loaded at startup and
others will be loaded when/if requested. For each branch to be loaded
the get_branch_func methoid is invoked with the id of the opened parent
stuck to the end.

Adding items:

if ($tree->has_feature(LOAD_ON_DEMAND))
{
	if ($row["parent"] == $rn)
	{
		$tree->add_item();
	}
}
else
{
	$tree->add_item();
};