+ automatic remoting by proxy objects for aw - cause it's just cool.
the way this works is: you set the is_remoted member of the class array in the ini file to point to the baseurl of the site this class should be remoted to, then ron 'make remoting' and it generated the proxy class for the remoted class. now, whenever you do a get_instance("remoted_class"); then you get the proxy object instance and when you make a call on the proxy object, it remotes it to the correct server and returns hte data. basically you just get classes and call functions on them and whena class needs to run on a special server, it will run there and your code won't even know.

if yuou call the class on the server it is supposed to be remoted to, you get a local call of course. the classes are not persistent yet, meaning that state is lost between successive calls to methods of the same object, that will come in the future sometime maybe

currently the site_list class is remoted to register.automatweb.com, meaning that if you do
$sl = get_instance("install/site_list");
echo $sl->orb_list();

you will get the correct site list, whatever the server your code is running on.

neat, eh?