+ auth servers support for aw v1.0
if you create an authentication configuration object and relate some authentication servers to it then aw will authenticate users against those, in the order you specified. for instance, you can say, first authenticate against this ldap server, then that and if all fails, local users

or you can say that authenticate against that ldap server and nothing else.

if someone logs in with an uid/pwd that is valid in the auth server but no such local user exists, and the auth config says so, the user object is automatically created and placed in a group, thus it gets privileges automatically.