+ now, not more than once a day, when an user logs in, a report is sent to register.automatweb.com with information about the site - the server is it running on, the code path, url and so on. thus the site list is always up to date.
the update is relatively secure - when the first update occurs, a secret key is created and stored both in the sites database and the registers database and after that all updates are encrypted with that secret. if the encryption does not matc, the update is rejected. thus, the security concern is that the attacker listens in during the initial key creation.

-