if aw objects datasource uses meta objects to give object_treeview_v2 the folders info, then it is possible to make some grouping according to these foldernames.
also multiple filters are handled more like i expect (like: filter1 AND filter2)

it is possible to group table rows. The best result comes if the table is sorted
by the field which will be used to make groups in table. If it is needed to sort
separateli under table groups, then secondary grouping options should be set and
it does the trick.