+ optimize object_list data reading - now when you do an object list, the queries go like this:
1) filter query that has the filter WHERE clause and reads oid,class_id,name, brother_of
if begin() is called:
2) for each class_id in the result, fetch all props for that type of objects

so if you just wanna do names() or count() then the functionality is exactly the same and just one query is made

but if you do begin() or arr() and there is just one type of object in the list, then you get away with just two querties for all the objects in the list.

still to do: refactor ds_mysql - it is really a mess by now. and there probably are still some more places that can be stuffed into one query