+ remove site_styles display for now. reason: performance. this thing gets called on *every* pageview for *every* site.
the total result of this was, that before aw could do 130kb/sec of output, after adding this message, it dropped to 80kb/sec.

so, when yuo put this back, make sure, that this gets called only when it needs to. for instance, require some ini setting or something to turn it on.