+ a solution for all your db upgrading problems:
if you add a db field, then in your class, define a method:

	function do_db_upgrade($tbl, $field, $q, $err)

in that method, do something like this:
		switch($field)
		{
			case "udef_ta1":
			case "udef_ta2":
			case "udef_ta3":
			case "udef_ta4":
			case "udef_ta5":
				$this->db_add_col($tbl, array(
					"name" => $field,
					"type" => "text"
				));
				return true;
		}
		return false;

basically, if tour method returns true , the db driver retries the failed query, if your method returns false, it is assumed that you can not upgrade to the correct version and an error message is generated.

the way this works is, when the db layer encounters an error, it tries to call the upgrade handler. this is done via error message parsing, which is definately not perfect, but should work for 90% of the cases.

so now, if there are any missing column errors, then it is your own damn fault.