+ major change in how cache works. now there are folders under pagecache: menu_area_cache storage_search storage_object_data html acl
and temp

now, when a certain cache needs to be cleared, it is renamed to the temp folder and a new one is created instead, so cache clear is almost instant.

now, for every user's session, once an hour the action maitenance::cache_update is requested, that clears all the files from the temp folder. this is requested in a way that the user will not notice the time it takes in their pageload.

basically, saving objects just got lots faster