Bug  #226023 Time:
If there is add_topic.tpl template on $site/templates/forum directory, then use this one, if not, then use the default webform.tpl template.

It is not the best solution in my opinion, as it only checks for site templates folder, but I can't think of any better solution to maintain backwards compatibility, cause I guess all the sites are using the webform.tpl to show the add_topic form and afaik there is some old add_topic.tpl present in code sides templates folder.